# copr: add simd for vector distance functions