--- a/components/tidb_query_datatype/Cargo.toml
+++ b/components/tidb_query_datatype/Cargo.toml
@@ -35,6 +35,7 @@ protobuf = "2"
 regex = "1.1"
 serde = "1.0"
 serde_json = "1.0"
+simsimd = "5.1.1"
 slog = { workspace = true }
 slog-global = { workspace = true }
 static_assertions = { version = "1.0", features = ["nightly"] }