--- a/components/tidb_query_datatype/benches/bench_vector_distance.rs
+++ b/components/tidb_query_datatype/benches/bench_vector_distance.rs
@@ -6,8 +6,8 @@ use tidb_query_datatype::codec::mysql::VectorFloat32Ref;
 fn bench_l1_distance_3d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.1, 2.2, 3.3];
     let vb: Vec<f32> = vec![1.1, 2.2, 3.3];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("l1_distance_3d", |b| {
         b.iter(|| {
@@ -19,8 +19,8 @@ fn bench_l1_distance_3d(c: &mut Criterion) {
 fn bench_l1_distance_784d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.0; 784];
     let vb: Vec<f32> = vec![1.0; 784];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("l1_distance_784d", |b| {
         b.iter(|| {
@@ -32,8 +32,8 @@ fn bench_l1_distance_784d(c: &mut Criterion) {
 fn bench_l2_squared_distance_3d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.1, 2.2, 3.3];
     let vb: Vec<f32> = vec![1.1, 2.2, 3.3];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("l2_squared_distance_3d", |b| {
         b.iter(|| {
@@ -49,8 +49,8 @@ fn bench_l2_squared_distance_3d(c: &mut Criterion) {
 fn bench_l2_squared_distance_784d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.0; 784];
     let vb: Vec<f32> = vec![1.0; 784];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("l2_squared_distance_784d", |b| {
         b.iter(|| {
@@ -66,8 +66,8 @@ fn bench_l2_squared_distance_784d(c: &mut Criterion) {
 fn bench_l2_distance_3d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.1, 2.2, 3.3];
     let vb: Vec<f32> = vec![1.1, 2.2, 3.3];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("l2_distance_3d", |b| {
         b.iter(|| {
@@ -79,8 +79,8 @@ fn bench_l2_distance_3d(c: &mut Criterion) {
 fn bench_l2_distance_784d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.0; 784];
     let vb: Vec<f32> = vec![1.0; 784];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("l2_distance_784d", |b| {
         b.iter(|| {
@@ -92,8 +92,8 @@ fn bench_l2_distance_784d(c: &mut Criterion) {
 fn bench_inner_product_3d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.1, 2.2, 3.3];
     let vb: Vec<f32> = vec![1.1, 2.2, 3.3];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("inner_product_3d", |b| {
         b.iter(|| {
@@ -105,8 +105,8 @@ fn bench_inner_product_3d(c: &mut Criterion) {
 fn bench_inner_product_784d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.0; 784];
     let vb: Vec<f32> = vec![1.0; 784];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("inner_product_784d", |b| {
         b.iter(|| {
@@ -118,8 +118,8 @@ fn bench_inner_product_784d(c: &mut Criterion) {
 fn bench_cosine_distance_3d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.1, 2.2, 3.3];
     let vb: Vec<f32> = vec![1.1, 2.2, 3.3];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("cosine_distance_3d", |b| {
         b.iter(|| {
@@ -135,8 +135,8 @@ fn bench_cosine_distance_3d(c: &mut Criterion) {
 fn bench_cosine_distance_784d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.0; 784];
     let vb: Vec<f32> = vec![1.0; 784];
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
-    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
+    let vec_vb = VectorFloat32Ref::from_f32(vb.as_slice());
 
     c.bench_function("cosine_distance_784d", |b| {
         b.iter(|| {
@@ -152,7 +152,7 @@ fn bench_cosine_distance_784d(c: &mut Criterion) {
 fn bench_l2_norm_3d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.1, 2.2, 3.3];
 
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
 
     c.bench_function("l2_norm_3d", |b| {
         b.iter(|| {
@@ -164,7 +164,7 @@ fn bench_l2_norm_3d(c: &mut Criterion) {
 fn bench_l2_norm_784d(c: &mut Criterion) {
     let va: Vec<f32> = vec![1.0; 784];
 
-    let vec_va = VectorFloat32Ref::from_f32(va.as_slice()).unwrap();
+    let vec_va = VectorFloat32Ref::from_f32(va.as_slice());
 
     c.bench_function("l2_norm_784d", |b| {
         b.iter(|| {