--- a/components/tidb_query_datatype/src/codec/data_type/chunked_vec_vector_float32.rs
+++ b/components/tidb_query_datatype/src/codec/data_type/chunked_vec_vector_float32.rs
@@ -126,3 +126,33 @@ impl<'a> UnsafeRefInto<&'static ChunkedVecVectorFloat32> for &'a ChunkedVecVecto
         std::mem::transmute(self)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_push_data_and_get() {
+        let mut chunked_vec = ChunkedVecVectorFloat32::with_capacity(5);
+
+        chunked_vec.push_data(VectorFloat32::copy_from_f32(&[1.1, 2.2, 3.3]));
+        chunked_vec.push_data(VectorFloat32::copy_from_f32(&[4.4, 5.5]));
+        // push a null value
+        chunked_vec.push_null();
+        chunked_vec.push_data(VectorFloat32::copy_from_f32(&[6.6, 7.7, 8.8, 9.9]));
+
+        assert_eq!(chunked_vec.len(), 4);
+
+        let vector1 = chunked_vec.get(0).unwrap().to_string();
+        assert_eq!(vector1, "[1.1,2.2,3.3]");
+
+        let vector2 = chunked_vec.get(1).unwrap().to_string();
+        assert_eq!(vector2, "[4.4,5.5]");
+
+        // check if null value is right
+        assert!(chunked_vec.get(2).is_none());
+
+        let vector3 = chunked_vec.get(3).unwrap().to_string();
+        assert_eq!(vector3, "[6.6,7.7,8.8,9.9]");
+    }
+}