<!--
Thank you for contributing to TiKV!

If you haven't already, please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document.

If you're unsure about anything, just ask; somebody should be along to answer within a day or two.

PR Title Format:
1. module [, module2, module3]: what's changed
2. *: what's changed
-->

### What is changed and how it works?
<!--
Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#linking-issues.
-->

Issue Number: Ref #17290 

Function Compute is replaced by simd：
- `l2_squared_distance`
- `inner_product`
- `cosine_distance`
- `l2_norm`   

<!--
You could use "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#format-of-the-commit-message.
-->
### What's Changed:
After optimization, the performance of low-dimensional vectors does not differ much, and the high-dimensional performance improvement is generally 200% +.The vector distance function calculation before using uncheck_index to get elements. Although in https://github.com/tikv/tikv/pull/17333 and https://github.com/tikv/tikv/pull/17334 we fixed memory alignment and improved bounds checking to optimize performance, we found that it still has room for optimization.

This pr defines the f32 parsing at the time of structured parsing of vectors, avoiding memory alignment problems when dealing with vectors. At the same time, the simd optimization function is used directly in the distance function to perform the computation, which improves the performance.

After optimization, the performance of low-dimensional vectors does not differ much, and **the high-dimensional performance improvement is generally 200%~300% +.**


- l2_squared_distance

| dimension | before       | after        | Speedup (x) |
|-----------|--------------|--------------|-------------|
| 3d        | 2.4573 ns    | 3.3492 ns    | 0.73        |
| 10d       | 4.3755 ns    | 2.6823 ns    | **1.63**    |
| 100d      | 33.421 ns    | 11.157 ns    | **3.00**    |
| 400d      | 206.12 ns    | 59.926 ns    | **3.44**    |
| 784d      | 487.49 ns    | 138.66 ns    | **3.52**    |
| 1000d     | 649.60 ns    | 192.56 ns    | **3.37**    |
| 4000d     | 2,838.3 ns   | 953.03 ns    | **2.98**    |

- l2_distance

| dimension | before       | after        | Speedup (x) |
|-----------|--------------|--------------|-------------|
| 3d        | 2.7286 ns    | 3.3500 ns    | 0.81        |
| 10d       | 4.6754 ns    | 3.0264 ns    | **1.55**    |
| 100d      | 34.306 ns    | 12.446 ns    | **2.76**    |
| 400d      | 206.32 ns    | 62.567 ns    | **3.30**    |
| 784d      | 491.73 ns    | 148.73 ns    | **3.31**    |
| 1000d     | 652.75 ns    | 194.32 ns    | **3.36**    |
| 4000d     | 2,870.2 ns   | 967.67 ns    | **2.97**    |

- inner_product

| dimension | before       | after        | Speedup (x) |
|-----------|--------------|--------------|-------------|
| 3d        | 2.4126 ns    | 3.5134 ns    | 0.69        |
| 10d       | 4.3488 ns    | 2.6561 ns    | **1.64**    |
| 100d      | 32.191 ns    | 10.209 ns    | **3.15**    |
| 400d      | 192.06 ns    | 57.238 ns    | **3.35**    |
| 784d      | 477.57 ns    | 129.00 ns    | **3.70**    |
| 1000d     | 634.12 ns    | 180.39 ns    | **3.52**    |
| 4000d     | 2,805.3 ns   | 978.09 ns    | **2.87**    |

- cosine_distance

| dimension | before       | after        | Speedup (x) |
|-----------|--------------|--------------|-------------|
| 3d        | 3.0262 ns    | 3.2580 ns    | 0.93        |
| 10d       | 6.5512 ns    | 3.4449 ns    | **1.90**    |
| 100d      | 60.630 ns    | 14.494 ns    | **4.18**    |
| 400d      | 278.40 ns    | 64.096 ns    | **4.34**    |
| 784d      | 563.94 ns    | 150.73 ns    | **3.74**    |
| 1000d     | 724.23 ns    | 217.23 ns    | **3.33**    |
| 4000d     | 2,953.6 ns   | 972.29 ns    | **3.04**    |

- l2_norm

| dimension      | before | after | Speedup (x) |
|-----------|---------------|---------------|-------------|
| 3d        | 1.8571 ns     | 3.6121 ns     | 0.51        |
| 10d       | 3.7841 ns     | 3.5989 ns     | **1.05**    |
| 100d      | 33.885 ns     | 12.432 ns     | **2.73**    |
| 400d      | 206.15 ns     | 58.894 ns     | **3.50**    |
| 784d      | 474.14 ns     | 134.61 ns     | **3.52**    |
| 1000d     | 619.50 ns     | 185.18 ns     | **3.35**    |
| 4000d     | 2,680.6 ns    | 977.16 ns     | **2.74**    |                        


### Related changes

- [ ] PR to update `pingcap/docs`/`pingcap/docs-cn`:
- [ ] Need to cherry-pick to the release branch

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

### Release note
<!-- 
Compatibility change, improvement, bugfix, and new feature need a release note.

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

If you don't think this PR needs a release note then fill it with None.
If this PR will be picked to release branch, then a release note is probably required.
-->

```release-note
copr: Improve vector distance performance
```