@wuhuizuo 2024-10-11
It introduced the building failure on linux/arm64 platform.